requirements.txt
streamlit
streamlit-drawable-canvas
streamlit-mic-recorder
requests
Pillow
gTTS